requirements.txt
click==8.1.3
Flask-MySQL==1.5.2
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.1.1
PyMySQL==1.0.2
PyYAML==6.0
setuptools==62.3.2
Werkzeug==2.1.2
wheel==0.37.1
zipp==3.8.0
